Move classes in p2p/sync & add progress updates tracker
