+++ p2p-sync/src/main/java/org/smartregister/p2p/model/dao/SenderTransferDao.java
@@ -5,8 +5,8 @@ import android.support.annotation.Nullable;
 import android.support.annotation.WorkerThread;
 
 import org.smartregister.p2p.model.DataType;
-import org.smartregister.p2p.sync.MultiMediaData;
-import org.smartregister.p2p.sync.JsonData;
+import org.smartregister.p2p.sync.data.MultiMediaData;
+import org.smartregister.p2p.sync.data.JsonData;
 
 import java.util.TreeSet;
