+++ p2p-sync/src/main/java/org/smartregister/p2p/presenter/P2PSenderPresenter.java
@@ -32,8 +32,8 @@ import org.smartregister.p2p.model.P2pReceivedHistory;
 import org.smartregister.p2p.sync.ConnectionLevel;
 import org.smartregister.p2p.sync.DiscoveredDevice;
 import org.smartregister.p2p.sync.ISenderSyncLifecycleCallback;
-import org.smartregister.p2p.sync.SyncPackageManifest;
-import org.smartregister.p2p.sync.SyncSenderHandler;
+import org.smartregister.p2p.sync.data.SyncPackageManifest;
+import org.smartregister.p2p.sync.handler.SyncSenderHandler;
 import org.smartregister.p2p.tasks.GenericAsyncTask;
 import org.smartregister.p2p.tasks.Tasker;
 import org.smartregister.p2p.util.Constants;
