--- p2p-sync/src/main/java/org/smartregister/p2p/sync/DataSendInterface.java
@@ -1,8 +1,0 @@
-package org.smartregister.p2p.sync;
-
-/**
- * Created by Ephraim Kigamba - ekigamba@ona.io on 28/03/2019
- */
-
-public interface DataSendInterface {
-}
