+++ p2p-sync/src/main/java/org/smartregister/p2p/sync/data/SyncPackageManifest.java
@@ -1,4 +1,4 @@
-package org.smartregister.p2p.sync;
+package org.smartregister.p2p.sync.data;
 
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
