+++ p2p-sync/src/test/java/org/smartregister/p2p/sync/SyncReceiverHandlerTest.java
@@ -30,6 +30,9 @@ import org.smartregister.p2p.model.dao.ReceiverTransferDao;
 import org.smartregister.p2p.model.dao.SenderTransferDao;
 import org.smartregister.p2p.shadows.ShadowAppDatabase;
 import org.smartregister.p2p.shadows.ShadowTasker;
+import org.smartregister.p2p.sync.data.ProcessedChunk;
+import org.smartregister.p2p.sync.data.SyncPackageManifest;
+import org.smartregister.p2p.sync.handler.SyncReceiverHandler;
 import org.smartregister.p2p.util.Constants;
 
 import java.io.ByteArrayInputStream;
