+++ sample/src/main/java/org/smartregister/p2p/sample/dao/SampleSenderDao.java
@@ -7,8 +7,8 @@ import android.support.annotation.Nullable;
 import org.json.JSONArray;
 import org.smartregister.p2p.model.DataType;
 import org.smartregister.p2p.model.dao.SenderTransferDao;
-import org.smartregister.p2p.sync.JsonData;
-import org.smartregister.p2p.sync.MultiMediaData;
+import org.smartregister.p2p.sync.data.JsonData;
+import org.smartregister.p2p.sync.data.MultiMediaData;
 import org.smartregister.p2p.tasks.GenericAsyncTask;
 
 import java.io.File;
